cross chain code commit

+++ module/nuls-cross-chain/cross-chain/src/main/java/io/nuls/crosschain/nuls/servive/impl/NulsCrossChainServiceImpl.java
@@ -12,7 +12,6 @@ import io.nuls.core.core.annotation.Autowired;
 import io.nuls.core.core.annotation.Component;
 import io.nuls.core.exception.NulsException;
 import io.nuls.core.log.Log;
-import io.nuls.core.model.ByteUtils;
 import io.nuls.core.model.StringUtils;
 import io.nuls.core.parse.JSONUtils;
 import io.nuls.core.rpc.util.NulsDateUtils;
@@ -24,18 +23,23 @@ import io.nuls.crosschain.nuls.constant.NulsCrossChainConstant;
 import io.nuls.crosschain.nuls.constant.NulsCrossChainErrorCode;
 import io.nuls.crosschain.nuls.model.bo.Chain;
 import io.nuls.crosschain.nuls.model.po.CtxStatusPO;
-import io.nuls.crosschain.nuls.rpc.call.*;
+import io.nuls.crosschain.nuls.rpc.call.AccountCall;
+import io.nuls.crosschain.nuls.rpc.call.BlockCall;
+import io.nuls.crosschain.nuls.rpc.call.ChainManagerCall;
+import io.nuls.crosschain.nuls.rpc.call.TransactionCall;
 import io.nuls.crosschain.nuls.srorage.*;
 import io.nuls.crosschain.nuls.utils.CommonUtil;
+import io.nuls.crosschain.nuls.utils.LoggerUtil;
 import io.nuls.crosschain.nuls.utils.TxUtil;
 import io.nuls.crosschain.nuls.utils.manager.ChainManager;
 import io.nuls.crosschain.nuls.utils.manager.CoinDataManager;
+import io.nuls.crosschain.nuls.utils.thread.CrossTxHandler;
 import io.nuls.crosschain.nuls.utils.validator.CrossTxValidator;
 
 import java.io.IOException;
 import java.util.*;
 
-import static io.nuls.crosschain.nuls.constant.NulsCrossChainConstant.*;
+import static io.nuls.crosschain.nuls.constant.NulsCrossChainConstant.CHAIN_ID_MIN;
 import static io.nuls.crosschain.nuls.constant.NulsCrossChainErrorCode.*;
 import static io.nuls.crosschain.nuls.constant.ParamConstant.*;
 
@@ -228,22 +232,23 @@ public class NulsCrossChainServiceImpl implements CrossChainService {
         if (chain == null) {
             return false;
         }
+        int syncStatus = BlockCall.getBlockStatus(chain);
         try {
             List<NulsHash> convertHashList = new ArrayList<>();
             List<NulsHash> ctxStatusList = new ArrayList<>();
             List<NulsHash> otherCtxList = new ArrayList<>();
             for (Transaction ctx : txs) {
                 NulsHash ctxHash = ctx.getHash();
-                if(verifiedCtxMap.get(chainId) != null){
+                if (verifiedCtxMap.get(chainId) != null) {
                     verifiedCtxMap.get(chainId).remove(ctxHash);
                 }
                 CoinData coinData = ctx.getCoinDataInstance();
                 int fromChainId = AddressTool.getChainIdByAddress(coinData.getFrom().get(0).getAddress());
                 int toChainId = AddressTool.getChainIdByAddress(coinData.getTo().get(0).getAddress());
-                if(chainId == toChainId){
+                if (chainId == toChainId) {
                     NulsHash convertHash = ctxHash;
                     if (!config.isMainNet()) {
-                        convertHash = TxUtil.friendConvertToMain(chain, ctx, null, TxType.CROSS_CHAIN).getHash();
+                        convertHash = TxUtil.friendConvertToMain(chain, ctx, TxType.CROSS_CHAIN).getHash();
                     }
                     if (!convertHashService.save(convertHash, ctxHash, chainId)) {
                         rollbackCtx(convertHashList, ctxStatusList, otherCtxList, chainId);
@@ -255,17 +260,17 @@ public class NulsCrossChainServiceImpl implements CrossChainService {
                         return false;
                     }
                     otherCtxList.add(convertHash);
-                }else{
+                } else {
                     if (!config.isMainNet()) {
-                        NulsHash convertHash = TxUtil.friendConvertToMain(chain, ctx, null, TxType.CROSS_CHAIN).getHash();
+                        NulsHash convertHash = TxUtil.friendConvertToMain(chain, ctx, TxType.CROSS_CHAIN).getHash();
                         if (!convertHashService.save(convertHash, ctxHash, chainId)) {
                             rollbackCtx(convertHashList, ctxStatusList, otherCtxList, chainId);
                             return false;
                         }
                         convertHashList.add(convertHash);
                     }
-                    //如果当前链不为发起链，则本链为主网中转链需清空签名在对交易对签名拜占庭
-                    if(chainId != fromChainId){
+                    //如果当前链不为发起链，则本链为主网中转链需清空签名在对交易对签名拜占庭，避免其他链向本链获取交易时失败
+                    if (chainId != fromChainId) {
                         ctx.setTransactionSignature(null);
                         if (!otherCtxService.save(ctxHash, ctx, chainId)) {
                             rollbackCtx(convertHashList, ctxStatusList, otherCtxList, chainId);
@@ -273,25 +278,28 @@ public class NulsCrossChainServiceImpl implements CrossChainService {
                         }
                         otherCtxList.add(ctxHash);
                     }
-                    CtxStatusPO ctxStatusPO = new CtxStatusPO(ctx, TxStatusEnum.UNCONFIRM.getStatus());
-                    if (!ctxStatusService.save(ctxHash, ctxStatusPO, chainId)) {
-                        rollbackCtx(convertHashList, ctxStatusList, otherCtxList, chainId);
-                        return false;
+                    //回滚重新打包，如果当前跨链交易已处理完成，则不需要重复处理
+                    CtxStatusPO ctxStatusPO = ctxStatusService.get(ctxHash, chainId);
+                    if(ctxStatusPO != null){
+                        if(ctxStatusPO.getStatus() == TxStatusEnum.CONFIRMED.getStatus()){
+                            chain.getLogger().info("该跨链转账交易之前已处理完成，不需重复处理：{}",ctxHash.toHex() );
+                            continue;
+                        }
                     }
                     ctxStatusList.add(ctxHash);
-                    chain.getLogger().debug("跨链交易提交完成，对跨链交易做拜占庭验证：{}",ctxHash.toHex());
+                    chain.getLogger().debug("跨链交易提交完成，对跨链转账交易做拜占庭验证：{}", ctxHash.toHex());
+                    //如果本链为主网通知跨链管理模块发起链与接收链资产变更
+                    if(config.isMainNet()){
+                        List<String> txStrList = new ArrayList<>();
+                        for (Transaction tx : txs) {
+                            txStrList.add(RPCUtil.encode(tx.serialize()));
+                        }
+                        String headerStr = RPCUtil.encode(blockHeader.serialize());
+                        ChainManagerCall.ctxAssetCirculateCommit(chainId, txStrList, headerStr);
+                    }
                     //发起拜占庭验证
-                    TxUtil.localCtxByzantine(ctx, chain);
+                    chain.getCrossTxThreadPool().execute(new CrossTxHandler(chain, ctx, syncStatus));
                 }
-            }
-            //如果本链为主网通知跨链管理模块发起链与接收链资产变更
-            if (config.isMainNet()) {
-                List<String> txStrList = new ArrayList<>();
-                for (Transaction tx : txs) {
-                    txStrList.add(RPCUtil.encode(tx.serialize()));
-                }
-                String headerStr = RPCUtil.encode(blockHeader.serialize());
-                ChainManagerCall.ctxAssetCirculateCommit(chainId, txStrList, headerStr);
             }
             chain.getLogger().info("高度：{} 的跨链交易提交完成\n", blockHeader.getHeight());
             return true;
@@ -315,28 +323,36 @@ public class NulsCrossChainServiceImpl implements CrossChainService {
                 int toChainId = AddressTool.getChainIdByAddress(coinData.getTo().get(0).getAddress());
                 NulsHash ctxHash = ctx.getHash();
                 if (chainId == fromChainId) {
-                    if (!ctxStatusService.delete(ctxHash,  chainId) || !convertHashService.delete(ctxHash, chainId)) {
+                    CtxStatusPO ctxStatusPO = ctxStatusService.get(ctxHash, chainId);
+                    if(ctxStatusPO != null){
+                        if(ctxStatusPO.getStatus() == TxStatusEnum.CONFIRMED.getStatus()){
+                            chain.getLogger().info("该跨链转账交易已处理完成，不需回滚：{}",ctxHash.toHex() );
+                            continue;
+                        }
+                    }
+                    if (!ctxStatusService.delete(ctxHash, chainId) || !convertHashService.delete(ctxHash, chainId)) {
                         return false;
                     }
                 } else if (chainId == toChainId) {
                     NulsHash convertHash = ctxHash;
                     if (!config.isMainNet() && ctx.getType() == config.getCrossCtxType()) {
-                        convertHash = TxUtil.friendConvertToMain(chain, ctx, null, TxType.CROSS_CHAIN).getHash();
+                        convertHash = TxUtil.friendConvertToMain(chain, ctx, TxType.CROSS_CHAIN).getHash();
                     }
                     if (!convertHashService.delete(convertHash, chainId)) {
                         return false;
                     }
                 } else {
+                    CtxStatusPO ctxStatusPO = ctxStatusService.get(ctxHash, chainId);
+                    if(ctxStatusPO != null){
+                        if(ctxStatusPO.getStatus() == TxStatusEnum.CONFIRMED.getStatus()){
+                            chain.getLogger().info("该跨链转账交易已处理完成，不需回滚：{}",ctxHash.toHex() );
+                            continue;
+                        }
+                    }
                     if (!ctxStatusService.delete(ctxHash, chainId) || !convertHashService.delete(ctxHash, chainId)) {
                         return false;
                     }
                 }
-            }
-            //需要被清理的跨链交易高度
-            long sendHeight = blockHeader.getHeight() + chain.getConfig().getSendHeight();
-            if (!sendHeightService.delete(sendHeight, chainId)) {
-                chain.getLogger().error("已广播跨链高度清除失败！\n\n");
-                return false;
             }
             //如果为主网通知跨链管理模块发起链与接收链资产变更
             if (config.isMainNet()) {
@@ -423,6 +439,7 @@ public class NulsCrossChainServiceImpl implements CrossChainService {
     @SuppressWarnings("unchecked")
     public Result getRegisteredChainInfoList(Map<String, Object> params) {
         Map<String, Object> result = new HashMap<>(2);
+        LoggerUtil.commonLog.info("------获取跨链资产信息---总长度：" + chainManager.getRegisteredCrossChainList().size());
         result.put(LIST, chainManager.getRegisteredCrossChainList());
         return Result.getSuccess(SUCCESS).setData(result);
     }
@@ -447,7 +464,7 @@ public class NulsCrossChainServiceImpl implements CrossChainService {
             return Result.getFailed(CHAIN_NOT_EXIST);
         }
         Map<String, Object> result = new HashMap<>(2);
-        result.put(VALUE, config.getByzantineRatio() * CommonUtil.getCurrentPackAddressList(chain).size()/ NulsCrossChainConstant.MAGIC_NUM_100);
+        result.put(VALUE, config.getByzantineRatio() * CommonUtil.getCurrentPackAddressList(chain).size() / NulsCrossChainConstant.MAGIC_NUM_100);
         return Result.getSuccess(SUCCESS).setData(result);
     }
